Splting DownloadZipAsync into DownloadAlbumAsync and DownloadPlaylistAsync

+++ RhythmiX.Service/Downloaders/ZipDownloader/ZipDownloader.cs
@@ -1,17 +1,22 @@
-﻿using System.IO.Compression;
+﻿using RhythmiX.Service.API.ResponseModels;
+using System.IO.Compression;
 using System.Net;
 
 namespace RhythmiX.Service.Downloaders.ZipDownloader
 {
     public static class ZipDownloader
     {
-        public static async Task DownloadZipAsync(string name, string zipPath, string imagePath, string folderPath)
+        public static async Task DownloadAlbumAsync(AlbumResponseModel response, string folderPath)
         {
+            string name = response.Results[0].Name;
+            string zipPath = response.Results[0].Zip;
+            string imagePath = response.Results[0].Image;
+
             Directory.CreateDirectory($"{folderPath}/{name}");
             Uri url = new Uri(zipPath);
             string zipSavePath = $"{folderPath}/{name}/{name}.zip";
 
-            if (File.Exists(folderPath))
+            if (File.Exists($"{folderPath}/{name}"))
                 return;
 
             using (WebClient client = new WebClient())
@@ -20,7 +25,28 @@ namespace RhythmiX.Service.Downloaders.ZipDownloader
                 await ImageDownloader.ImageDownloader.DownoladImageAsync(name, imagePath, folderPath);
             }
 
-            ZipFile.ExtractToDirectory(zipSavePath, folderPath);
+            ZipFile.ExtractToDirectory(zipSavePath, $"{folderPath}/{name}");
+            File.Delete(zipSavePath);
+        }
+
+        public static async Task DownloadPlaylistAsync(PlaylistResponseModel response, string folderPath)
+        {
+            string name = response.Results[0].Name;
+            string zipPath = response.Results[0].Zip;
+
+            Directory.CreateDirectory($"{folderPath}/{name}");
+            Uri url = new Uri(zipPath);
+            string zipSavePath = $"{folderPath}/{name}/{name}.zip";
+
+            if (File.Exists($"{folderPath}/{name}"))
+                return;
+
+            using (WebClient client = new WebClient())
+            {
+                await client.DownloadFileTaskAsync(url, zipSavePath);
+            }
+
+            ZipFile.ExtractToDirectory(zipSavePath, $"{folderPath}/{name}");
             File.Delete(zipSavePath);
         }
     }
